Implement support for heartbeat, ping, pong in a WS connection
Kraken specifies that they send a heartbeat about every second,
and that application level pings can be sent to confirm that the
connection is still alive if a heartbeat is missed.

In this commit, we make the WS connection listen to these
heartbeats and check periodically if we have received messages
recently -- if not then we send a ping, and if this isn't answered
in a second then we close the stream.

This should fix issues where connectivity issues lead to undetected
states where the book stops getting updates.